fix(tests): move response-transformer integration tests to plugins.rs, drop duplicates from gateway.rs
The 192-test block in gateway.rs duplicated tests already structured into
per-feature files under crates/barbacane-test/tests/ (since dbb7822). Only
the 6 response-transformer integration tests were new — moved them to
plugins.rs alongside the existing request-transformer tests.
